refact: refact result set
