+++ java/openmldb-jdbc/src/test/java/com/_4paradigm/openmldb/jdbc/SQLRouterSmokeTest.java
@@ -24,6 +24,7 @@ import com._4paradigm.openmldb.sdk.Column;
 import com._4paradigm.openmldb.sdk.Schema;
 import com._4paradigm.openmldb.sdk.SdkOption;
 import com._4paradigm.openmldb.sdk.SqlExecutor;
+import com._4paradigm.openmldb.sdk.impl.SQLResultSet;
 import com._4paradigm.openmldb.sdk.impl.SqlClusterExecutor;
 
 import org.testng.Assert;
@@ -128,7 +129,7 @@ public class SQLRouterSmokeTest {
 
             // select
             String select1 = "select * from tsql1010;";
-            com._4paradigm.openmldb.jdbc.SQLResultSet rs1 = (com._4paradigm.openmldb.jdbc.SQLResultSet) router
+            SQLResultSet rs1 = (SQLResultSet) router
                     .executeSQL(dbname, select1);
 
             Assert.assertEquals(2, rs1.GetInternalSchema().getColumnList().size());
@@ -150,7 +151,7 @@ public class SQLRouterSmokeTest {
             rs1.close();
 
             String select2 = "select col1 from tsql1010;";
-            com._4paradigm.openmldb.jdbc.SQLResultSet rs2 = (com._4paradigm.openmldb.jdbc.SQLResultSet) router
+            SQLResultSet rs2 = (SQLResultSet) router
                     .executeSQL(dbname, select2);
             Assert.assertEquals(1, rs2.GetInternalSchema().size());
             Assert.assertEquals(Types.BIGINT, rs2.GetInternalSchema().getColumnType(0));
@@ -165,7 +166,7 @@ public class SQLRouterSmokeTest {
             rs2.close();
 
             String select3 = "select col2 from tsql1010;";
-            com._4paradigm.openmldb.jdbc.SQLResultSet rs3 = (com._4paradigm.openmldb.jdbc.SQLResultSet) router
+            SQLResultSet rs3 = (SQLResultSet) router
                     .executeSQL(dbname, select3);
             Assert.assertEquals(1, rs3.GetInternalSchema().size());
             Assert.assertEquals(Types.VARCHAR, rs3.GetInternalSchema().getColumnType(0));
@@ -185,7 +186,7 @@ public class SQLRouterSmokeTest {
             {
                 query_statement.setString(1, "hi");
                 query_statement.setLong(2, 1003);
-                com._4paradigm.openmldb.jdbc.SQLResultSet rs4 = (com._4paradigm.openmldb.jdbc.SQLResultSet) query_statement
+                SQLResultSet rs4 = (SQLResultSet) query_statement
                         .executeQuery();
                 Assert.assertEquals(2, rs4.GetInternalSchema().size());
                 Assert.assertEquals(Types.BIGINT, rs4.GetInternalSchema().getColumnType(0));
@@ -200,7 +201,7 @@ public class SQLRouterSmokeTest {
             {
                 query_statement.setString(1, "hi");
                 query_statement.setLong(2, 1002);
-                com._4paradigm.openmldb.jdbc.SQLResultSet rs4 = (com._4paradigm.openmldb.jdbc.SQLResultSet) query_statement
+                SQLResultSet rs4 = (SQLResultSet) query_statement
                         .executeQuery();
                 Assert.assertEquals(2, rs4.GetInternalSchema().size());
                 Assert.assertEquals(Types.BIGINT, rs4.GetInternalSchema().getColumnType(0));
@@ -212,7 +213,7 @@ public class SQLRouterSmokeTest {
             {
                 query_statement.setString(1, "world");
                 query_statement.setLong(2, 1003);
-                com._4paradigm.openmldb.jdbc.SQLResultSet rs4 = (com._4paradigm.openmldb.jdbc.SQLResultSet) query_statement
+                SQLResultSet rs4 = (SQLResultSet) query_statement
                         .executeQuery();
                 Assert.assertEquals(2, rs4.GetInternalSchema().size());
                 Assert.assertEquals(Types.BIGINT, rs4.GetInternalSchema().getColumnType(0));
@@ -227,7 +228,7 @@ public class SQLRouterSmokeTest {
             {
                 query_statement.setString(1, "hello");
                 query_statement.setLong(2, 1003);
-                com._4paradigm.openmldb.jdbc.SQLResultSet rs4 = (com._4paradigm.openmldb.jdbc.SQLResultSet) query_statement
+                SQLResultSet rs4 = (SQLResultSet) query_statement
                         .executeQuery();
                 Assert.assertEquals(2, rs4.GetInternalSchema().size());
                 Assert.assertEquals(Types.BIGINT, rs4.GetInternalSchema().getColumnType(0));
@@ -242,7 +243,7 @@ public class SQLRouterSmokeTest {
             {
                 query_statement.setString(1, "word");
                 query_statement.setLong(2, 1003);
-                com._4paradigm.openmldb.jdbc.SQLResultSet rs4 = (com._4paradigm.openmldb.jdbc.SQLResultSet) query_statement
+                SQLResultSet rs4 = (SQLResultSet) query_statement
                         .executeQuery();
                 Assert.assertEquals(2, rs4.GetInternalSchema().size());
                 Assert.assertEquals(Types.BIGINT, rs4.GetInternalSchema().getColumnType(0));
@@ -284,7 +285,7 @@ public class SQLRouterSmokeTest {
             // missing 2nd parameter
             {
                 query_statement.setString(1, "hi");
-                com._4paradigm.openmldb.jdbc.SQLResultSet rs4 = (com._4paradigm.openmldb.jdbc.SQLResultSet) query_statement
+                SQLResultSet rs4 = (SQLResultSet) query_statement
                         .executeQuery();
                 Assert.fail("executeQuery is expected to throw exception");
                 rs4.close();
@@ -426,7 +427,7 @@ public class SQLRouterSmokeTest {
             Assert.assertTrue(ok);
             // select
             String select1 = "select * from tsql1010;";
-            com._4paradigm.openmldb.jdbc.SQLResultSet rs1 = (com._4paradigm.openmldb.jdbc.SQLResultSet) router
+            SQLResultSet rs1 = (SQLResultSet) router
                     .executeSQL(dbname, select1);
             Assert.assertEquals(5, rs1.GetInternalSchema().size());
             Assert.assertEquals(Types.BIGINT, rs1.GetInternalSchema().getColumnType(0));
@@ -451,7 +452,7 @@ public class SQLRouterSmokeTest {
             rs1.close();
 
             String select2 = "select col1 from tsql1010;";
-            com._4paradigm.openmldb.jdbc.SQLResultSet rs2 = (com._4paradigm.openmldb.jdbc.SQLResultSet) router
+            SQLResultSet rs2 = (SQLResultSet) router
                     .executeSQL(dbname, select2);
             Assert.assertEquals(1, rs2.GetInternalSchema().size());
             Assert.assertEquals(Types.BIGINT, rs2.GetInternalSchema().getColumnType(0));
@@ -536,7 +537,7 @@ public class SQLRouterSmokeTest {
             impl.executeBatch();
             Assert.assertTrue(ok);
             String select1 = "select * from tsql1010;";
-            com._4paradigm.openmldb.jdbc.SQLResultSet rs1 = (com._4paradigm.openmldb.jdbc.SQLResultSet) router
+            SQLResultSet rs1 = (SQLResultSet) router
                     .executeSQL(dbname, select1);
             Assert.assertEquals(6, rs1.GetInternalSchema().size());
             rs1.close();
@@ -583,7 +584,7 @@ public class SQLRouterSmokeTest {
             Assert.assertEquals(result, expected);
 
             String select2 = "select * from tsql1010;";
-            com._4paradigm.openmldb.jdbc.SQLResultSet rs2 = (com._4paradigm.openmldb.jdbc.SQLResultSet) router
+            SQLResultSet rs2 = (SQLResultSet) router
                     .executeSQL(dbname, select1);
             Assert.assertEquals(6, rs2.GetInternalSchema().size());
             int recordCnt = 0;
